+++ lib/widget/challenge_menu.dart
@@ -1,0 +1,7 @@
+class ChallengeMenu {
+  String name;
+  int type;
+  dynamic data;
+
+  ChallengeMenu({this.name, this.type, this.data});
+}
