+++ pubspec.yaml
@@ -38,6 +38,7 @@ dependencies:
   firebase_messaging: 8.0.0-dev.8
   flutter_local_notifications: ^3.0.1+3
   permission_handler: ^4.4.0
+  flutter_app_badger: ^1.1.2
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
